handle | as a middle delimiter
fixes #6

The `bracketed_expression` grammar now checks if the 'middle' bit
contains another bracketed expression whose left delimiter is a
left/right bracket symbol. If so, it interprets that as a 'middle'
delimiter followed by the right delimiter.

This also adds the `|:` and `:|` symbols for left and right vertical
delimiters, when you want to explicitly specify them. The grammar is
still ambiguous, particularly in matrices.